Refactor: rename test package to memorchess
