+++ composeApp/src/androidTest/kotlin/proj/memorchess/axl/TestSettings.kt
@@ -1,4 +1,4 @@
-package proj.akichess.axl
+package proj.memorchess.axl
 
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithTag
